Add files via upload
It is a website built using Python which is developed to retrieve patient’s details easily. In this website doctor directly collects all the information of the patient regarding the disease like patient name, age, gender, B.P, weight blood test, urine test place and the description of the disease.